Add three extra tags
These tags are used by other similar plugins, so it'd be good to have this plugin show up along with them when people search these tags. For example: https://wordpress.org/plugins/tags/admin-message

I've included the plurals as well, as it seems to be what lots of plugins do.